+++ R/available-packages.R
@@ -155,7 +155,8 @@ renv_available_packages_success <- function(db, url) {
 renv_available_packages_entry <- function(package,
                                           type   = "source",
                                           filter = NULL,
-                                          quiet  = FALSE)
+                                          quiet  = FALSE,
+                                          prefer = NULL)
 {
 
   # if filter is a string, treat it as an explicit version requirement
@@ -175,7 +176,17 @@ renv_available_packages_entry <- function(package,
     entries[ordered[[1]], ]
   }
 
+  # read available packages
   dbs <- renv_available_packages(type = type, quiet = quiet)
+
+  # if a preferred repository is marked and available, prefer using that
+  if (length(prefer) == 1L && prefer %in% names(dbs)) {
+    idx <- match(prefer, names(dbs))
+    ord <- c(idx, setdiff(seq_along(dbs), idx))
+    dbs <- dbs[ord]
+  }
+
+  # iterate through repositories, and find first matching
   for (i in seq_along(dbs)) {
 
     db <- dbs[[i]]